welcoming_agent_1/pyproject.toml
[project]
name = "welcoming-agent-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]
